Create StringValidation.py
Python has built-in string validation methods for basic data. It can check if a string is composed of alphabetical characters, alphanumeric characters, digits, etc. 

str.isalnum() - This method checks if all the characters of a string are alphanumeric (a-z, A-Z and 0-9).

str.isalpha() - This method checks if all the characters of a string are alphabetical (a-z and A-Z).

str.digit() - This method checks if all the characters of a string are digits (0-9).

str.lower() - This method checks if all the characters of a string are lowercase characters (a-z).

str.upper() - This method checks if all the characters of a string are uppercase characters (A-Z).

Question:- You are given a string 'S'.
Your task is to find out if the string 'S' contains: alphanumeric characters, alphabetical characters, digits, lowercase and uppercase characters.

Output Format:- In the first line, print True if  has any alphanumeric characters. Otherwise, print False.
In the second line, print True if 'S' has any alphabetical characters. Otherwise, print False.
In the third line, print True if 'S' has any digits. Otherwise, print False.
In the fourth line, print True if 'S' has any lowercase characters. Otherwise, print False.
In the fifth line, print True if 'S' has any uppercase characters. Otherwise, print False.